requirements.txt
apify-client==1.7.1
supabase==2.4.0
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1
requests==2.31.0
python-dotenv==1.0.1
schedule==1.2.1
anthropic==0.25.0
pytz==2024.1
